# docs/swagger.yaml
basePath: /
definitions:
...
  CreateSocialMediaResponse:
    properties:
      created_at:
        example: "2022-10-10T11:52:28.431369Z"
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Instagram
        type: string
      url:
        example: https://instagram.com
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  DeleteResponse:
    description: DeleteResponse Model
    properties:
      message:
        example: your account has been successfully deleted
        type: string
    type: object
  LoginRequest:
    properties:
      email:
        example: abdianrizky11@gmail.com
        maxLength: 32
        minLength: 6
        type: string
      password:
        example: bcrypt hashed password
        maxLength: 32
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  LoginResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        type: string
    type: object
...
  RegisterRequest:
    properties:
      age:
        example: 18
        minimum: 8
        type: integer
      email:
        example: abdianrizky11@gmail.com
        type: string
      password:
        example: bcrypt hashed password
        maxLength: 32
        minLength: 6
        type: string
      username:
        example: adnsm
        type: string
    required:
    - age
    - email
    - password
    - username
    type: object
  RegisterResponse:
    properties:
      age:
        example: 18
        type: integer
      email:
        example: abdianrizky11@gmail.com
        type: string
      id:
        example: 1
        type: integer
      username:
        example: adnsm
        type: string
    type: object
  Social:
    description: Social Media Model
    properties:
      created_at:
        example: 2017-07-22'T'16:28:55.444
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Facebook
        type: string
      social_media_url:
        example: https://images.unsplash.com
        type: string
      updated_at:
        example: 2017-07-22'T'16:28:55.444
        type: string
      user:
        $ref: '#/definitions/User'
      user_id:
        example: 1
        type: integer
    type: object
  SocialMediaRequest:
    description: Social Media Request Model
    properties:
      name:
        example: Facebook
        type: string
      social_media_url:
        example: https://images.unsplash.com
        type: string
    required:
    - name
    - social_media_url
    type: object
...
  UpdateSocialMediaResponse:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: Instagram
        type: string
      social_media_url:
        example: https://instagram.com
        type: string
      updated_at:
        example: "2022-10-10T11:52:28.431369Z"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  UpdateUserRequest:
    properties:
      email:
        example: abdianrizky11@gmail.com
        maxLength: 32
        minLength: 6
        type: string
      username:
        example: adnsm
        type: string
    required:
    - email
    - username
    type: object
  UpdateUserResponse:
    properties:
      age:
        example: 18
        type: integer
      email:
        example: abdianrizky11@gmail.com
        type: string
      id:
        example: 1
        type: integer
      updated_at:
        example: "2022-10-10T11:52:28.431369Z"
        type: string
      username:
        example: adnsm
        type: string
    type: object
  User:
    description: User Model
    properties:
      email:
        example: admin@localhost
        type: string
      id:
        example: 1
        type: integer
      username:
        example: admin
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: dev.abdianrizky@gmail.com
    name: Abdian Rizky
    url: https://linktr.ee/adonism2k
  description: Ini adalah dokumentasi API Contract untuk tugas Final Project Hactiv8.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Final Project API
  version: "1.0"
paths:
...
  /socialmedias/:
    get:
      consumes:
      - application/json
      description: Get the current Social Media data
      parameters:
      - description: Authorization
        in: header
        name: auth
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            items:
              $ref: '#/definitions/Social'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get All Social Media
      tags:
      - Social Media
    post:
      consumes:
      - application/json
      description: Create a Social Media
      parameters:
      - description: Authorization
        in: header
        name: auth
        required: true
        type: string
      - description: Create Social Media Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SocialMediaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/CreateSocialMediaResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Social Media
      tags:
      - Social Media
  /socialmedias/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Social Media
      parameters:
      - description: Social Media ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/DeleteResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Social Media
      tags:
      - Social Media
    put:
      consumes:
      - application/json
      description: Update the current Social Media data
      parameters:
      - description: Authorization
        in: header
        name: auth
        required: true
        type: string
      - description: Social Media ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Social Media Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SocialMediaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/UpdateSocialMediaResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Social Media
      tags:
      - Social Media
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete the current user account
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/DeleteResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete User
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update the current user data
      parameters:
      - description: Authorization
        in: header
        name: auth
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/UpdateUserResponse'
      security:
      - ApiKeyAuth: []
      summary: Update User
      tags:
      - User
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns JWT token
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/LoginResponse'
      summary: Login
      tags:
      - User
  /users/register:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns JWT token
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/RegisterResponse'
      summary: Register
      tags:
      - User
securityDefinitions:
  "":
    description: Bearer Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"